// Codigo/tathy-festa-web-app/src/types/Cliente.ts
export interface Cliente {
  id: number;
  nomeCompleto: string;
  cpf: string;
  identidade: string;
  nomeAniversariante: string;
  idadeAniversariante: string;
  endereco: string;
  telefone: string;
  email: string;
  instagram: string;
  enderecoEvento: string;
  dataEvento: Date;
  horaEvento: string;
  formaPagamento: string;
  comoConheceu: string;
}